Clean up (#410)
* Moving ImportExportConfig to the Config folder
Removing unneeded castbar subclasses

* More clean up

* :)
