+++ DelvUI/Config/ConfigurationManager.cs
@@ -1,5 +1,4 @@
 using Dalamud.Plugin;
-using DelvUI.Config.Attributes;
 using DelvUI.Config.Tree;
 using DelvUI.Helpers;
 using DelvUI.Interface;
@@ -9,11 +8,9 @@ using DelvUI.Interface.StatusEffects;
 using ImGuiNET;
 using ImGuiScene;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
 using System;
 using System.IO;
 using System.IO.Compression;
-using System.Linq;
 using System.Reflection;
 using System.Text;
 
@@ -187,12 +184,10 @@ namespace DelvUI.Config
 
         public static string ExportBaseNode(BaseNode baseNode)
         {
-            //SkipNonPortableContractResolver resolver = new SkipNonPortableContractResolver();
             JsonSerializerSettings settings = new JsonSerializerSettings
             {
                 TypeNameAssemblyFormatHandling = TypeNameAssemblyFormatHandling.Simple,
                 TypeNameHandling = TypeNameHandling.Objects
-                //ContractResolver = new SkipNonPortableContractResolver()
             };
             var jsonString = JsonConvert.SerializeObject(baseNode, Formatting.Indented, settings);
             ImGui.SetClipboardText(jsonString);
@@ -204,7 +199,7 @@ namespace DelvUI.Config
             // see comments on ConfigPageNode's Load
             MethodInfo methodInfo = typeof(ConfigurationManager).GetMethod("LoadImportString");
             MethodInfo function = methodInfo.MakeGenericMethod(configPageNode.ConfigObject.GetType());
-            PluginConfigObject importedConfigObject = (PluginConfigObject)function.Invoke(ConfigurationManager.GetInstance(), new object[] { importString });
+            PluginConfigObject importedConfigObject = (PluginConfigObject)function.Invoke(GetInstance(), new object[] { importString });
 
             if (importedConfigObject != null)
             {
@@ -212,8 +207,8 @@ namespace DelvUI.Config
                 // update the tree
                 configPageNode.ConfigObject = importedConfigObject;
                 // but also update the dictionary
-                ConfigurationManager.GetInstance().ConfigBaseNode.configObjectsMap[configPageNode.ConfigObject.GetType()] = configPageNode.ConfigObject;
-                ConfigurationManager.GetInstance().SaveConfigurations();
+                GetInstance().ConfigBaseNode.configObjectsMap[configPageNode.ConfigObject.GetType()] = configPageNode.ConfigObject;
+                GetInstance().SaveConfigurations();
                 _instance.ResetEvent(_instance, null);
             }
             else
