--- DelvUI/Interface/ImportExportConfig.cs
@@ -1,86 +1,0 @@
-﻿using Dalamud.Plugin;
-using DelvUI.Config;
-using DelvUI.Config.Attributes;
-using ImGuiNET;
-using Newtonsoft.Json;
-using System;
-using System.Numerics;
-
-namespace DelvUI.Interface
-{
-    [Portable(false)]
-    [Section("Import/Export")]
-    [SubSection("General", 0)]
-    public class ImportExportConfig : PluginConfigObject
-    {
-        private string _importString = "";
-        private string _exportString = "";
-
-        public new static ImportExportConfig DefaultConfig() { return new ImportExportConfig(); }
-
-        [ManualDraw]
-        public void DrawFullImportExport()
-        {
-            uint maxLength = 100000;
-            ImGui.BeginChild("importpane", new Vector2(0, ImGui.GetWindowHeight() / 6), false, ImGuiWindowFlags.NoScrollbar | ImGuiWindowFlags.NoScrollWithMouse);
-            {
-                ImGui.Text("Import string:");
-                ImGui.InputText("", ref _importString, maxLength);
-
-                if (ImGui.Button("Import configuration") && _importString != "")
-                {
-                    string[] importStrings = _importString.Trim().Split(new string[] { "|" }, StringSplitOptions.RemoveEmptyEntries);
-                    ConfigurationManager.LoadTotalConfiguration(importStrings);
-                    //ConfigurationManager.ImportBaseNode(_importString);
-                }
-
-                ImGui.SameLine();
-
-                if (ImGui.Button("Paste from clipboard"))
-                {
-                    try
-                    {
-                        _importString = ImGui.GetClipboardText();
-                    }
-                    catch (Exception ex)
-                    {
-                        PluginLog.Log("Could not get clipboard text:\n" + ex.StackTrace);
-                    }
-
-                }
-            }
-
-            ImGui.EndChild();
-
-            ImGui.BeginChild("exportpane", new Vector2(0, ImGui.GetWindowHeight() / 6), false, ImGuiWindowFlags.NoScrollbar | ImGuiWindowFlags.NoScrollWithMouse);
-
-            {
-                ImGui.Text("Export string:");
-                ImGui.InputText("", ref _exportString, maxLength, ImGuiInputTextFlags.ReadOnly);
-
-                if (ImGui.Button("Export configuration"))
-                {
-                    _exportString = ConfigurationManager.GetInstance().ConfigBaseNode.GetBase64String();
-                    //_exportString = ConfigurationManager.ExportBaseNode(ConfigurationManager.GetInstance().ConfigBaseNode);
-                }
-
-                ImGui.SameLine();
-
-                if (ImGui.Button("Copy to clipboard") && _exportString != "")
-                {
-                    try
-                    {
-                        ImGui.SetClipboardText(_exportString);
-                    }
-                    catch (Exception ex)
-                    {
-                        PluginLog.Log("Could not set clipboard text:\n" + ex.StackTrace);
-                    }
-
-                }
-            }
-
-            ImGui.EndChild();
-        }
-    }
-}
